requirements.txt
streamlit
pandas
numpy
gspread
google-auth
requests
python-dotenv
matplotlib
seaborn
tabulate
